chore(hf_gpt2): Re-structures hf_gpt2-related files.
