+++ archai/nlp/models/hf_gpt2/onnx_hf_gpt2.py
@@ -9,19 +9,23 @@ from typing import Any, Dict
 import torch
 from onnxruntime.transformers.onnx_model_gpt2 import Gpt2OnnxModel as HfGPT2OnnxModel
 
-from archai.nlp.models.config_base import BATCH_SIZE, SEQ_LEN, OnnxConfig
+from archai.nlp.models.config_base import OnnxConfig
 
 
 class HfGPT2OnnxConfig(OnnxConfig):
-    def __init__(self, model_config: str) -> None:
-        super().__init__(model_config)
+    """Huggingface's Open AI GPT-2 ONNX-based configuration.
 
-        self.config['past_key_values'] = 2
-        self.config['model_type'] = 'gpt2'
+    """
+
+    def __init__(self, model_config: Dict[str, Any]) -> None:
+        model_config['past_key_values'] = 2
+        model_config['model_type'] = 'gpt2'
+
+        super().__init__(model_config)
 
     @property
     def mockups(self) -> Dict[str, Any]:
         return {
-            'input_ids': torch.randint(0, self.config['n_token'], (BATCH_SIZE, SEQ_LEN)),
-            'past_key_values': tuple([torch.zeros(self.config['past_key_values'], BATCH_SIZE, self.config['n_head'], SEQ_LEN, self.config['d_head']) for _ in range(self.config['n_layer'])])
+            'input_ids': torch.randint(0, self.config.n_token, (self.batch_size, self.seq_len)),
+            'past_key_values': tuple([torch.zeros(self.config.past_key_values, self.batch_size, self.config.n_head, self.seq_len, self.config.d_head) for _ in range(self.config.n_layer)])
         }
